change task 2 polymorphism

+++ src/polymorphism/task2/TaxService.java
@@ -1,7 +1,9 @@
 package polymorphism.task2;
 
+import java.math.BigDecimal;
+
 class TaxService {
-    public void payOut(double taxAmount) {
+    public void payOut(BigDecimal taxAmount) {
         System.out.format("Уплачен налог в размере %.2f%n", taxAmount);
     }
 }
